labelWidth derives from NSSize so make it a CGFloat
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3947 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/AddressView/ADPersonPropertyView.h
@@ -99,8 +99,8 @@ typedef enum {
 - (void) setDisplaysLabel: (BOOL) yesno;
 - (BOOL) displaysLabel;
 
-- (void) setMaxLabelWidth: (NSInteger) width;
-- (NSInteger) maxLabelWidth;
+- (void) setMaxLabelWidth: (CGFloat) width;
+- (CGFloat) maxLabelWidth;
 
 - (NSFont*) font;
 - (void) setFont: (NSFont*) font;
